Refactored method of reading from Twitter and some error checking
